reporting: Updated reporting component to work data sample based.
The reporting component can now detect if new data has arrived on ports or not.
This requires the latest changes on rtt-2.0-mainline which add checks during updating
of data sources.
AutoTrigger has been removed since it does not make sense anymore. snapshot() will always
behave as expected: it causes a new line to be reported with all current values.
A NullSample string property has been added to be outputted if no new data is present.