deployer: only load libraries ending in .so,.dll,... during dir scan
This was necessary to let the symlinks to their work, ie
point to the most recent version, unless the user overrides
with a specific version (like foo.so.1.2.3) in which case
that version is loaded. So the .so|.dll|.dylib extension
restriction only applies when scanning a whole directory
for components (RTT_COMPONENT_PATH etc).

Signed-off-by: Peter Soetens <peter@thesourceworks.com>