cmake: fix linking error when cross-compiling.
Executables linked with ocl-deployment or ocl-taskbrowser weren't linking correctly when cross-compiling. The executables weren't explicitely linked with the dependencies of ocl-deployment and ocl-taskbrowser. Therefore, the linker was using the shared libraries RPATH property to find out the path of the plugins. This isn't good because the encoded RPATH is using a full path and the linker seems to look at it relative to the provided sysroot. One workaround it to force the executables to link with dependencies of ocl-deployment and ocl-taskbrowser.