lua: Fix statistics in TLSF when memory region is added.
When a new memory region is added, free_ex is called on it to mark it
as free. This however messes up the statistics, as the new size is
then subtracted from tlsf->used_size.

This hack temporarily stores the correct size and restores it after
free_ex has been called.