netcdf: performance and serialization fixes
In some places, whole copies were made of properties/property bags.
The ncreport example was broken ( wrong activity setting )
The field names for structs or sequences were not prefixed, ending
all up in the top level namespace.
Decomposition was previously turned off, but this also rules out structs.
so now we just decompose (Decompose=1), which also makes this usable for
kst(2) which can't read arrays of doubles anyway.

This patch fixes all that.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>