Extend deployer to find and scan optional .rtt_component_path file
This file is a list of relative paths to be added to the RTT_COMPONENT_PATH
environment variable.  This makes it possible to ship code that builds and runs
without forcing the user to manually set up environment environment variables
(Or rely on another software like ROS to do the forcing). This file might be
added to the root of a git repository, for example.