logging: Support direct connection to Category log port
Unit test code frequently requires direct access to a category's
log port, to be able to syphon events off. This supports direct
connection without needing to know the category port's name, which
is usually the method to locate a particularly port within a
component. The problem with category ports is that their name
is a reverse-DNS-order translation of the category name, and so
can be tricky to figure out.