bin: Stop deployment script processing after the first failure
Processing of all scripts occured even when one or more scripts failed
to load. In layered systems, if an earlier layer failed to deploy then
later layers could have serious issues without the underlying
functionality that they require.

This fix ensures that once a script fails to deploy, then processing
of scripts stops. If applicable, the TaskBrowser or CORBA server is
still started to allow user inspection to identify the problem.

This also unifies the default settings; previously the cdeployer and
deployer defaulted to true, while the deployer-corba defaulted to
false.