Avoiding Apple's readline in favor of Homebrew's
The readline/editline implementation provided by Apple is completely
broken and makes compilation impossible. As a consequence, it is
better to interrupt building and show a clear error message.

This means that it must be handled as a special case: just prefixing
Homebrew's paths would allow CMake to continue searching in the standard
Unix locations, hence finding the broken readline.