cmake: Fix dependency after log4cpp renaming
If the non-Orocos log4cpp isn't installed, then pkgconfig won't find
this component, ocl-logging, as the 'log4cpp' transitive dependency
won't be found as it's now called 'orocos-log4cpp'. This means that
this component is unusable by any dependent package.

The fix is to rename the library to match the changes in the 'log4cpp'
toolchain package.