bin: make utility scripts shell-agnostic, use exec and added OROCOS_LAUNCH_PREFIX
Shell scripts are configured by cmake's configure_file() command to avoid the need to
find the installation path based on the script's location, using CMAKE_INSTALL_PREFIX
directly instead. This works in any shell and not only with bash.

With the exec shell command the shell will be replaced by the Orocos process instead
of running it as child.

The OROCOS_LAUNCH_PREFIX simplifies debugging when the deployer or other executables
are run from a script or by an external tool like roslaunch.

Example:
  OROCOS_LAUNCH_PREFIX="gdb -ex run --args" ./user_script_that_executes_the_deployer_with_arguments.sh

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>