logging: Support connecting a log port to a new category
When setting the level or additivity, if the category of interest
doesn't yet exist then it is created before its level/additivity is
set. But when connecting the log port to an appender, the category
must exist first. This assymetry can make configuration of complex
applications difficult, and the behaviour is not clear.

The fix is to allow connecting a log port of a non-existent category
by first creating the category. This makes the behaviour consistent
and clear.

Also added comments to the level and additivity processing, to indicate
that a category will be created if necessary.