logging: Optimize memory allocations by removing stack-based LoggingEvent
Make the "event" a class member, rather than a stack instance, to reduce the
number of real-time memory allocations. The strings inside \a event will
eventually allocate to the largest string seen in the system.

WARNING: processEvents() is not multi-thread safe now! (though it should
never have been called from multiple-threads in the first place)