deployment: Fix passing component name to all activities
Without this the thread name is set to "FileDescriptorActivity", etc, which
is useless when multiple of each activity type are present in the deployment.

The fix is to pass the component name to the activity class constructor, the
same as with the other Activity derived classes.

WARNING PeriodicActivity does not support a name, so nothing to do there.