timer: Name the underlying Timer activity/thread
Currently the parent component and the internal catcher thread have
the same name, which is confusing.

Use the component name + ".Timer". This way the component might get
"MyTimer", and the internal catcher thread becomes "MyTimer.Timer"
to separate it from the parent component.

NB: Thread names have a 15 character limit!